Write tests (#250)
* Minor changes to services

* Minor change to model validations and added factory for schedules table

* Made slight changes to model valdiations

* Added model tests

* Added controller tests

* Minor change to delete schedule on article update service and added tests

* Added tests for services and modified services

* Added another model validation, and modified some tests

+++ app/services/delete_schedule_service.rb
@@ -1,14 +1,13 @@
 # frozen_string_literal: true
 
 class DeleteScheduleService
-  attr_reader :id, :current_user
+  attr_reader :id
 
-  def initialize(id, current_user)
+  def initialize(id)
     @id = id
-    @current_user = current_user
   end
 
-  def process
+  def process!
     delete_schedule_service
   end
 
@@ -16,10 +15,11 @@ class DeleteScheduleService
 
     def delete_schedule_service
       schedule = ArticleStatusSchedule.find(id)
-      article = current_user.articles.find(schedule.article_id)
-      schedules_for_article = article.article_status_schedules.all.order(scheduled_time: :asc)
+      article = Article.find(schedule.article_id)
+      schedules_for_article = article.article_status_schedules
+        .where(schedule_status: :pending).order(scheduled_time: :asc)
       duplicate_schedule = schedules_for_article.find_by("scheduled_time > ?", schedule.scheduled_time)
-      if !duplicate_schedule.nil?
+      if schedule.schedule_status == "pending" && !duplicate_schedule.nil?
         duplicate_schedule.destroy!
       end
       schedule.destroy!
